feat: bookEntity에 있던 구독자 컬럼 따로 분리 구현, 쿼리문 수정

+++ src/main/java/com/project/novel/dto/BookDto.java
@@ -8,14 +8,14 @@ import java.time.LocalDateTime;
 public class BookDto {
     private Long id;
     private String bookName;
-    private int subscribes;
+    private Long subscribes;
     private String author;
     private Long hit;
     private Long likes;
     private LocalDateTime modifiedDate;
 
 
-    public BookDto(Long id, String bookName, int subscribes, String author, Long hit, Long likes, LocalDateTime modifiedDate) {
+    public BookDto(Long id, String bookName, Long subscribes, String author, Long hit, Long likes, LocalDateTime modifiedDate) {
         this.id = id;
         this.bookName = bookName;
         this.subscribes = subscribes;
